requirements.txt
torch
torchvision
matplotlib
wandb
Pillow
numpy